# docs/api/api.yaml
openapi: "3.0.0"
info:
  title: "Wutsi Order API"
  description: API for managing orders
  version: "1.0.0"
servers:
  - url: https://wutsi-order-server-test.herokuapp.com
    description: Sandbox
  - url: https://wutsi-order-server-prod.herokuapp.com
    description: Production
paths:
...
  /v1/orders/search:
    post:
      operationId: "search-orders"
      description: Search Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchOrderRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchOrderResponse'
        403:
          description: The caller is not authorized to call this endpoint
      tags:
        - Order
      security:
        - api_key: [ 'order-read' ]
...
  /v1/orders/{id}/status:
    post:
      operationId: "change-status"
      description: Change an order status
      parameters:
        - in: path
          name: id
          description: ID of the order
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeStatusRequest'
      responses:
        200:
          description: Success
        403:
          description: The caller is not authorized to call this endpoint
      tags:
        - Order
      security:
        - api_key: [ 'order-manage' ]
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    OrderSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID of the order
        merchantId:
          type: integer
          format: int64
          description: ID of the merchant
        accountId:
          type: integer
          format: int64
          description: ID of the customer
        reservationId:
          type: integer
          format: int64
          description: ID of the reservation associated with Order
          nullable: true
        status:
          type: string
          description: Status of the order
        paymentStatus:
          type: string
          description: Status of the payment
        subTotalPrice:
          type: number
          format: double
          description: Sub Total, which is the total price of the products.
        deliveryFees:
          type: number
          format: double
          description: Delivery fees
        savingsAmount:
          type: number
          format: double
          description: Savings amount. Savings are computed by summing for each product, the difference between unit price vs comparable-price
        totalPrice:
          type: number
          format: double
          description: Total price to charge to the customer
        totalPaid:
          type: number
          format: double
          description: Total amount paid by the customer
        currency:
          type: string
          description: Currency
        created:
          type: string
          format: date-time
          description: Creation datetime
        updated:
          type: string
          format: date-time
          description: Last update
    SearchOrderRequest:
      type: object
      properties:
        merchantId:
          type: integer
          format: int64
          description: ID of the merchant
          nullable: true
        accountId:
          type: integer
          format: int64
          description: ID of the customer
          nullable: true
        status:
          type: array
          items:
            type: string
          description: Status of the order
        createdFrom:
          type: string
          format: date-time
          description: Creation datetime start date
          nullable: true
        createdTo:
          type: string
          format: date-time
          description: Creation datetime start date
          nullable: true
        limit:
          type: integer
          default: 30
        offset:
          type: integer
          default: 0
    SearchOrderResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderSummary"
...
    ChangeStatusRequest:
      type: object
      properties:
        status:
          type: string
          description: New status
        reason:
          type: string
          description: Reason of the change
          nullable: true
        comment:
          type: string
          description: Comment
          nullable: true